requirements.txt
django-bootstrap3<=5.4.0